Roster Trunk: - Removed the rest of the menu config items - Added config for login, search, and locale boxes in header under Display settings - Added display of realmstatus, not in a good position yet - Changed realmstatus timestamp to a real unix timestamp - Merged Realm Status display and type into one config item - Remoave menuleft, menuright, makepanel, make realmstatus code from menu - Added makerealmstatus code in functions.lib.php - Fixed include path for guildinfo bargraph images - Re-added footer text for class and image lists in text mode for guildinfo - Added new parameters for makelink() -- $anchor: (bool) allows you to disable the anchor tag in the link -- $ext: (string) allows you to change the url extension for the link, default is html - Changed tooltip building from string concatenate to an array, should remove extra br tags - Added back to roster button after upgrade

+++ trunk/lib/dbal/structure/mysql_structure.sql
@@ -735,7 +735,7 @@ CREATE TABLE `renprefix_realmstatus` (
   `servertype` varchar(20) NOT NULL default '',
   `serverstatus` varchar(20) NOT NULL default '',
   `serverpop` varchar(20) NOT NULL default '',
-  `timestamp` tinyint(2) NOT NULL default '0',
+  `timestamp` int(11) NOT NULL default '0',
   UNIQUE KEY `server_name` (`server_name`)
 ) ENGINE=MyISAM DEFAULT CHARSET=utf8;
